Refactor table styling documentation and update CSS variables for improved clarity and consistency
