+++ docs/docs/components/table/table_dynamic_styling.md
@@ -1,5 +1,5 @@
 ---
-sidebar_position: 45
+sidebar_position: 21
 title: Dynamic Styling
 slug: styling
 ---
@@ -8,12 +8,19 @@ The `Table` component lets you style specific rows or cells using custom part na
 
 ## Row styling
 
-The `setRowPartProvider()` method assigns style names to entire rows based on the data item they contain. This lets you highlight full rows that meet specific conditions—for example, alternating background colors for even rows.
+The `setRowPartProvider()` method assigns part names to entire rows based on the data item they contain. This lets you highlight full rows that meet specific conditions—for example, alternating background colors for even rows.
 
 These style names can be targeted using the `::part()` selector in your CSS.
 
+:::tip Shadow parts
+The `::part()` selector is a special CSS feature that allows you to style elements inside a component's shadow DOM—as long as those elements expose a `part` attribute. This is especially useful for styling internal parts of webforJ components, like rows or cells in a table.
+
+For more on how shadow parts work and how to define and target them, see the [Styling](../../styling/shadow-parts) section.
+:::
+
+
 <ComponentDemo 
-path='http://localhost:8080/webforj/tablerowstyling?' 
+path='/webforj/tablerowstyling?' 
 javaE='https://raw.githubusercontent.com/webforj/webforj-documentation/refs/heads/main/src/main/java/com/webforj/samples/views/table/TableRowStylingView.java'
 height='300px'
 />
@@ -25,7 +32,7 @@ The `setCellPartProvider()` method styles individual cells based on both the dat
 Like row parts, cell parts are defined by a name and targeted using the `::part()` selector.
 
 <ComponentDemo 
-path='http://localhost:8080/webforj/tablecellstyling?' 
+path='/webforj/tablecellstyling?' 
 javaE='https://raw.githubusercontent.com/webforj/webforj-documentation/refs/heads/main/src/main/java/com/webforj/samples/views/table/TableColumnPinningView.java'
 height='300px'
 />
@@ -33,10 +40,10 @@ height='300px'
 
 ## Reacting to data updates
 
-If your app modifies data programmatically, such as updating a user's age, the table will reflect the new styles automatically after committing the changes using `table.getRepository().commit(item)`.
+If your app modifies data programmatically, such as updating a user's age, the table will automatically re-evaluate and reapply any associated row or cell styles once the updated item is committed in the repository 
 
 <ComponentDemo 
-path='http://localhost:8080/webforj/tabledynamicstyling?' 
+path='/webforj/tabledynamicstyling?' 
 javaE='https://raw.githubusercontent.com/webforj/webforj-documentation/refs/heads/main/src/main/java/com/webforj/samples/views/table/TableDynamicStylingView.java'
 height='475px'
 />
