+++ src/main/java/com/webforj/samples/views/table/Person.java
@@ -1,0 +1,37 @@
+package com.webforj.samples.views.table;
+
+class Person {
+  private String name;
+  private int age;
+  private String city;
+
+  public Person(String name, int age, String city) {
+    this.name = name;
+    this.age = age;
+    this.city = city;
+  }
+
+  public String getName() {
+    return name;
+  }
+
+  public void setName(String name) {
+    this.name = name;
+  }
+
+  public int getAge() {
+    return age;
+  }
+
+  public void setAge(int age) {
+    this.age = age;
+  }
+
+  public String getCity() {
+    return city;
+  }
+
+  public void setCity(String city) {
+    this.city = city;
+  }
+}
