+++ src/main/java/com/webforj/samples/views/table/TableDynamicStylingView.java
@@ -16,24 +16,24 @@ import com.webforj.router.annotation.Route;
 @Route
 @InlineStyleSheet(/* css */ """
   dwc-table::part(row-even) {
-    background-color: var(--dwc-color-default);
+    background-color: var(--dwc-color-gray-alt);
   }
 
   dwc-table::part(cell-senior) {
-    background-color: var(--dwc-color-success);
-    color: var(--dwc-color-on-success-text);
+    background-color: var(--dwc-color-success-alt);
+    color: var(--dwc-color-success-text);
   }
 
   dwc-table::part(cell-junior) {
-    background-color: var(--dwc-color-danger);
-    color: var(--dwc-color-on-danger-text);
+    background-color: var(--dwc-color-danger-alt);
+    color: var(--dwc-color-danger-text);
   }
 """)
 public class TableDynamicStylingView extends Composite<FlexLayout> {
 
   private final FlexLayout self = getBoundComponent();
   private final Table<Person> table = new Table<>();
-  private final Button update = new Button("Toggle Alice Age", ButtonTheme.GRAY);
+  private final Button update = new Button("Increase Alice Age", ButtonTheme.GRAY);
   private boolean toggleAge = false;
 
   public TableDynamicStylingView() {
@@ -43,7 +43,7 @@ public class TableDynamicStylingView extends Composite<FlexLayout> {
 
     List<Person> data = List.of(
         new Person("Alice", 28, "New York"),
-        new Person("Alice", 32, "Chicago"),
+        new Person("Tom", 32, "Chicago"),
         new Person("Bob", 28, "Chicago"),
         new Person("Bob", 35, "New York"),
         new Person("Charlie", 25, "Los Angeles"),
@@ -89,41 +89,5 @@ public class TableDynamicStylingView extends Composite<FlexLayout> {
     });
 
     self.add(update, table);
-  }
-
-  public static class Person {
-    private String name;
-    private int age;
-    private String city;
-
-    public Person(String name, int age, String city) {
-      this.name = name;
-      this.age = age;
-      this.city = city;
-    }
-
-    public String getName() {
-      return name;
-    }
-
-    public void setName(String name) {
-      this.name = name;
-    }
-
-    public int getAge() {
-      return age;
-    }
-
-    public void setAge(int age) {
-      this.age = age;
-    }
-
-    public String getCity() {
-      return city;
-    }
-
-    public void setCity(String city) {
-      this.city = city;
-    }
   }
 }
